fix error in question3

+++ Question3.java
@@ -11,7 +11,7 @@ public class Question3
      */
     // System.out.println("Enter a whole number: "); 
     Scanner in = new Scanner(System.in);
-    double result = Math.pow(in.nextInt(),2);
+    int result = Math.pow(in.nextInt(),2);
     System.out.println(result);
   }
 }
